Refactor app.js by pulling routes into their own separate module
